android-version/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
bcrypt==4.1.2
psutil==5.9.8
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
celery==5.3.6
redis==5.0.1
